Lion_maze/requirements.txt
numpy==1.17.4
pillow>=6.2.0
termcolor==1.1.0

pyenchant~=3.2.0